Fix transaction logging in BankAccount history
The log member variable in the BankAccount model would almost never
work, always returning null.

Java always treats strings as immutable, so no matter what the
append or concat function would do, it would not affect the member
variable (A new local copy of log is made and then changed, but the
original unaffected).

This was fixed with the addition of the mutable StringBuffer class,
which will now allow for proper concatenation of events in the log
variable