Refactor transferTo test, create getLogMsg helper
In the tests, parseLog(bankAccount.getLog(), index) is everywhere, but
the only variable that sometimes changes is the index. (It only changes
in transferTo where we need to check the target Account's log to verify
the transaction).

I wrote a quick, dirty, and sinful hack that I will fix later, but not
right now