Remove redundant overflow unit tests for BankAccount
I'm not sure how the client will handle overflows. If Exceptions are
used, then users with big balances for bank accounts would crash/error on
initialization. If BigDouble/Big Equivalent is used, then the program
would run fine until it ran out of memory to store that value (not ideal
on a phone or mobile device).

In short I think the best way to handle this is to not implement this
feature yet until it is necessary to focus on overflows later on after
the rest of the application is made. And then it should probably be
handled server side anyways, and balance will store only a shorthand of
the real account balance.