Implement transaction searching by keywords
In order to get the logs to work, the BankAccountTests were just parsing
the log at arbitrary intervals. Searching by regex allows for an easier
way to test the logs, and also more fine-grained searching. An
alternative would be to add specific tags or perhaps use a csv file to
store the history.

A bug was also fixed with the transferTo_Acct1000 test.