Refactor AESCipher to use private Data obj
AESCipher was always passing around a Data object, or variable of the
Data object in almost every one of its methods. So in order to increase
clarity it has become a variable of AESCipher and all the methods have
been reworked to remove the Data data parameter.

DataFactory has been deleted as all the Data generation is within
AESCipher.