Fix bug with hack where Overlay did not expand
The Overlay was not set in column 0, row 0 and was also not
expanding across the columns and rows to fill them. The bug
was fixed by removing the GridPane that was in the Dashboard
fxml file.

Ideally the Dashboard would look differently from the overlay
with some nice welcome messages, but for now it will be the same
as the overlay.